+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added support for subscripts
 - Type collections `types.based`, `types.implementing` and `types.inheriting` now return non-optional array. If no types found, empty array will be returned. 
 This is a breaking change for template code like this:
+- You can now pass additional arguments one by one, i.e. `--args arg1=value1 --args arg2 --args arg3=value3`
 
  ```swift
 <% for type in (types.implementing["SomeProtocol"] ?? []) { %>
@@ -28,6 +29,7 @@ This is a breaking change for template code like this:
 - JS exceptions no more override syntax errors in JS templates 
 - Accessing unknown property on `types` now results in a better error than `undefined is not an object` in JS templates
 - Fixed issue in AutoMockable, where generated non-optional variables wouldn't meet protocol's requirements. For this purpose, underlying variable was introduced
+- Fixed `inline:auto` not inserting code if Sourcery is run with cache enabled #467
 
 ## 0.10.1